dataset_metadata_ingest/requirements.txt
beautifulsoup4>=4.12.3
httpx>=0.27.2
huggingface_hub>=0.29.0
kaggle>=1.7.4.5
lxml>=5.3.0
psycopg[binary]>=3.2.3
PyYAML>=6.0.2
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
